requirements.txt
freezegun==0.3.12
Django==3.0.2
django-appconf==1.0.3
django-axes==4.5.4
django-cors-headers==3.2.1
django-extensions==2.2.5
django-filter==2.2.0
django-redis==4.10.0
django-storages==1.6.6
django-ipware==2.1.0
django-guardian==2.3.0
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.9
djangorestframework==3.11.0
djangorestframework-guardian==0.3.0
djangorestframework-jwt==1.11.0
drf-autodocs==0.4.4
drf-yasg
factory-boy==2.12.0
gunicorn==20.0.4
PyJWT==1.7.1
python-jwt==3.2.5
swagger-spec-validator==2.4.3
validate-email==1.3
psycopg2-binary
Pillow==5.2.0
redis
